pyproject.toml
[project]
name = "analyze-fin"
version = "0.1.0"
description = "Philippine Personal Finance Tracker - Parse bank statements, categorize transactions, generate insights"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "arjay"}
]
keywords = ["finance", "banking", "philippines", "gcash", "bpi", "maya"]

dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pdfplumber>=0.11",
    "typer>=0.12",
    "rich>=13.0",
    "jinja2>=3.1",
    "plotly>=5.18",
    "pydantic>=2.5",
    "numpy>=2.4.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
analyze-fin = "analyze_fin.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analyze_fin"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (Typer pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["analyze_fin"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
